% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{capacity_cdf}
\alias{capacity_cdf}
\title{Capacity cdf}
\usage{
capacity_cdf(x, ...)
}
\arguments{
\item{x}{An S7 capacity, routing, or recharge component, as required by the generic.}

\item{...}{Method arguments: see Details for the component-specific calling convention.}
}
\value{
Numeric values or a component result as described in Details.
}
\description{
Call capacity_cdf(x, c) for nonnegative point capacities c in mm.
}
\examples{
capacity_cdf(FlodeParetoCapacity(0, 140, 0.4), c = c(0, 70, 140))

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
