% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{CapacityDistribution}
\alias{CapacityDistribution}
\title{Compatibility alias for FlodeCapacityDistribution}
\usage{
CapacityDistribution()
}
\description{
Compatibility alias for FlodeCapacityDistribution
}
\seealso{
\link{FlodeCapacityDistribution}
}
