% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{cubic_routing}
\alias{cubic_routing}
\title{Cubic routing}
\usage{
cubic_routing(kb = 10000, solver = "adaptive")
}
\arguments{
\item{kb}{Positive reciprocal rate coefficient: hour mm^(m-1).}

\item{solver}{Character scalar: adaptive or smith. The Smith approximation requires m = 3.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
cubic_routing(kb = 10000)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
