% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{DemandRecharge}
\alias{DemandRecharge}
\title{Compatibility alias for FlodeDemandRecharge}
\usage{
DemandRecharge(alpha = integer(0), beta = integer(0), qsat = integer(0))
}
\arguments{
\item{alpha}{Demand deficit threshold in \verb{(0,1]}, or fast-path split fraction in \verb{[0,1]}.}

\item{beta}{Positive exponent of groundwater demand.}

\item{qsat}{Positive saturation recharge rate in mm/hour. Requires positive implied Sgmax.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
FlodeDemandRecharge(0.5, 1, 1)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
