% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating.R
\name{error_updater}
\alias{error_updater}
\title{Error updater}
\usage{
error_updater(ar = c(0.7), ma = numeric(), scale = "additive", offset = 0)
}
\arguments{
\item{ar}{Finite vector of R-convention autoregressive coefficients (negative phi in Moore).}

\item{ma}{Finite vector of moving-average coefficients in R convention.}

\item{scale}{Error scale: additive or log.}

\item{offset}{Nonnegative offset in cubic metres per second for log error updating.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
error_updater(ar = 0.7)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
