% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{ExponentialCapacity}
\alias{ExponentialCapacity}
\title{Compatibility alias for FlodeExponentialCapacity}
\usage{
ExponentialCapacity(mean = integer(0))
}
\arguments{
\item{mean}{Positive mean point capacity in mm.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
FlodeExponentialCapacity(100)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
