% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{ExponentialRouting}
\alias{ExponentialRouting}
\title{Compatibility alias for FlodeExponentialRouting}
\usage{
ExponentialRouting(a = integer(0), gamma = integer(0))
}
\arguments{
\item{a}{Positive exponential-routing slope, per mm.}

\item{gamma}{Finite exponential-routing log-flow intercept, flow measured in mm/hour.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
FlodeExponentialRouting(0.05, -5)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
