% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{FlodeCapacityDistribution}
\alias{FlodeCapacityDistribution}
\title{Create FlodeCapacityDistribution}
\usage{
FlodeCapacityDistribution()
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\details{
This is an abstract extension class. Construct a concrete subclass.
}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
