% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{FlodeCascadeRouting}
\alias{FlodeCascadeRouting}
\title{Create FlodeCascadeRouting}
\usage{
FlodeCascadeRouting(k1 = integer(0), k2 = integer(0))
}
\arguments{
\item{k1}{Positive first linear-reservoir time constant, hours.}

\item{k2}{Positive second linear-reservoir time constant, hours.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
FlodeCascadeRouting(2, 4)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
