% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{FlodePdmModel}
\alias{FlodePdmModel}
\title{Create FlodePdmModel}
\usage{
FlodePdmModel(
  distribution = FlodeCapacityDistribution(),
  recharge = FlodeRecharge(),
  surface = FlodeRouting(),
  groundwater = FlodeRouting(),
  be = integer(0),
  fc = integer(0),
  td = integer(0),
  qc = integer(0),
  area_km2 = integer(0),
  dt = integer(0)
)
}
\arguments{
\item{distribution}{An S7 capacity distribution object.}

\item{recharge}{An S7 recharge component.}

\item{surface}{An S7 fast routing component, or initial routing storage vector in mm for state
functions.}

\item{groundwater}{An S7 slow routing component, or initial storage vector in mm for state
functions.}

\item{be}{Positive exponent in the actual evaporation function.}

\item{fc}{Nonnegative dimensionless rainfall correction factor.}

\item{td}{Nonnegative common outlet delay in hours.}

\item{qc}{Finite signed external constant flow in cubic metres per second.}

\item{area_km2}{Positive catchment area in square kilometres.}

\item{dt}{Positive model interval in hours. Internal paper equations use hours.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
pdm_model()

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
