% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recharge.R
\name{FlodeRecharge}
\alias{FlodeRecharge}
\title{Create FlodeRecharge}
\usage{
FlodeRecharge()
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\details{
This is an abstract extension class. Construct a concrete subclass.
}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
