% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recharge.R
\name{FlodeSplitRecharge}
\alias{FlodeSplitRecharge}
\title{Create FlodeSplitRecharge}
\usage{
FlodeSplitRecharge(alpha = integer(0))
}
\arguments{
\item{alpha}{Demand deficit threshold in \verb{(0,1]}, or fast-path split fraction in \verb{[0,1]}.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
FlodeSplitRecharge(0.7)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
