% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{FlodeTriangularCapacity}
\alias{FlodeTriangularCapacity}
\title{Create FlodeTriangularCapacity}
\usage{
FlodeTriangularCapacity(cmin = integer(0), cmax = integer(0))
}
\arguments{
\item{cmin}{Minimum point storage capacity in mm; nonnegative and below cmax.}

\item{cmax}{Maximum point storage capacity in mm; strictly above cmin.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
FlodeTriangularCapacity(0, 200)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
