% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{linear_routing}
\alias{linear_routing}
\title{Linear routing}
\usage{
linear_routing(kb = 24)
}
\arguments{
\item{kb}{Positive reciprocal rate coefficient: hour mm^(m-1).}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
linear_routing(kb = 24)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
