% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdm_config.R
\name{load_pdm_model}
\alias{load_pdm_model}
\title{Load pdm model}
\usage{
load_pdm_model(path)
}
\arguments{
\item{path}{Character scalar identifying a YAML file. Executable expressions are rejected.}
}
\value{
A validated FlodePdmModel.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
load_pdm_model(system.file('config', 'standard.yml', package = 'pdmS7'))

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
