% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{LognormalCapacity}
\alias{LognormalCapacity}
\title{Compatibility alias for FlodeLognormalCapacity}
\usage{
LognormalCapacity(meanlog = integer(0), sdlog = integer(0))
}
\arguments{
\item{meanlog}{Finite mean of log point capacity (capacity measured in mm).}

\item{sdlog}{Positive standard deviation of log point capacity.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
FlodeLognormalCapacity(log(100) - 0.5, 1)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
