% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{parameter_catalog}
\alias{parameter_catalog}
\title{Parameter catalog}
\usage{
parameter_catalog()
}
\value{
A data.frame of parameters, units, domains, examples, equation references, and
provenance.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
head(parameter_catalog())

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
