% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{ParetoCapacity}
\alias{ParetoCapacity}
\title{Compatibility alias for FlodeParetoCapacity}
\usage{
ParetoCapacity(cmin = integer(0), cmax = integer(0), b = integer(0))
}
\arguments{
\item{cmin}{Minimum point storage capacity in mm; nonnegative and below cmax.}

\item{cmax}{Maximum point storage capacity in mm; strictly above cmin.}

\item{b}{Nonnegative Pareto shape exponent; zero represents a point mass at cmax.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
FlodeParetoCapacity(0, 140, 0.4)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
