% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{pdm_from_config}
\alias{pdm_from_config}
\title{Pdm from config}
\usage{
pdm_from_config(config)
}
\arguments{
\item{config}{Named model configuration list; unknown keys are rejected.}
}
\value{
A validated FlodePdmModel.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
pdm_from_config(pdm_presets('standard'))

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
