% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{quadratic_routing}
\alias{quadratic_routing}
\title{Quadratic routing}
\usage{
quadratic_routing(kb = 1000)
}
\arguments{
\item{kb}{Positive reciprocal rate coefficient: hour mm^(m-1).}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
quadratic_routing(kb = 1000)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
