% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdm_config.R
\name{read_pdm_config}
\alias{read_pdm_config}
\title{Read pdm config}
\usage{
read_pdm_config(path)
}
\arguments{
\item{path}{Character scalar identifying a YAML file. Executable expressions are rejected.}
}
\value{
A validated named configuration list.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
read_pdm_config(system.file('config', 'standard.yml', package = 'pdmS7'))

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
