% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{routing_flow}
\alias{routing_flow}
\title{Routing flow}
\usage{
routing_flow(x, ...)
}
\arguments{
\item{x}{An S7 capacity, routing, or recharge component, as required by the generic.}

\item{...}{Method arguments: see Details for the component-specific calling convention.}
}
\value{
Numeric values or a component result as described in Details.
}
\description{
Call routing_flow(x, storage), returning endpoint discharge in mm/hour.
}
\examples{
routing_flow(linear_routing(24), storage = 24)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
