% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_2.R
\name{run_pdm}
\alias{run_pdm}
\title{Run a forcing sequence}
\usage{
run_pdm(model, forcing, state = initial_state(model))
}
\arguments{
\item{model}{A FlodePdmModel compatible with the supplied state.}

\item{forcing}{Data frame with rain and pet interval depths in mm and optional POSIXct time.}

\item{state}{A compatible FlodePdmState; NULL in sim_pdm initializes the model.}
}
\value{
A list containing output, final state, initial state, model and water balance.
}
\description{
Run a forcing sequence
}
