% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_2.R
\name{soil_step}
\alias{soil_step}
\title{Soil step}
\usage{
soil_step(distribution, storage, net_depth)
}
\arguments{
\item{distribution}{An S7 capacity distribution object.}

\item{storage}{Basin soil storage in mm for soil_step.}

\item{net_depth}{Signed net interval depth in mm, after evaporation and recharge.}
}
\value{
A list with storage and runoff, both in mm.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
soil_step(FlodeParetoCapacity(0, 140, 0.4), storage = 50, net_depth = 10)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
