% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating.R
\name{state_updater}
\alias{state_updater}
\title{State updater}
\usage{
state_updater(
  scheme = "proportional",
  gb = 1,
  gs = 1,
  gg = 0,
  beta1 = 10,
  beta2 = 1.1,
  soil_proportional = TRUE
)
}
\arguments{
\item{scheme}{State correction rule: proportional, super_proportional, or non_proportional.}

\item{gb}{Nonnegative dimensionless baseflow correction gain.}

\item{gs}{Nonnegative dimensionless surface-flow correction gain.}

\item{gg}{Nonnegative soil-storage correction gain in hours; zero disables soil correction.}

\item{beta1}{Positive surface-flow weighting in super-proportional correction.}

\item{beta2}{Groundwater weighting, at least one, in super-proportional correction.}

\item{soil_proportional}{Logical scalar; use the baseflow proportion for soil correction if
TRUE.}
}
\value{
An S7 component object.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
state_updater(scheme = 'super_proportional')

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
