% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_part2.R
\name{transfer_coefficients}
\alias{transfer_coefficients}
\title{Transfer coefficients}
\usage{
transfer_coefficients(k1, k2, dt = 1)
}
\arguments{
\item{k1}{Positive first linear-reservoir time constant, hours.}

\item{k2}{Positive second linear-reservoir time constant, hours.}

\item{dt}{Positive model interval in hours. Internal paper equations use hours.}
}
\value{
Named numeric vector delta1, delta2, omega0, omega1 for endpoint recursion.
}
\description{
Component of the configurable Moore (2007) rainfall-runoff framework.
}
\examples{
transfer_coefficients(k1 = 2, k2 = 4, dt = 0.25)

}
\references{
Moore (2007), doi:10.5194/hess-11-483-2007.
}
